+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](https://semver.org/)
 
+## [4.0.0] 2025-07-04
+
+### Changed
+
+- update template to latest develop (rdflib 7, cmem-plugin-base 4.12.0)
+ 
 ## [3.0.1] 2025-03-13
 
 ### Fixed